Adding test for write_csv()
This is a fairly simple test but is highlighting a windows based newline error.
On windows the csv files get extra lines added "\r\r\n" which puts a gap between each row. not a problem on the linux server but annoying for testing. There is an easy fix to be added in the next commit.

Testing
Ran pytest - 1 test (this new csv test) fails. - as explained above