Adding - Basic tests for dcat
This is a fairly simple set of tests for dcat.py. It doesn't not cover everything in detail but covers the broad brush strokes - to be coloured in more later.

Some minor changes have been made to dact.py to allow it to be tested without contacting the web. This allows for repeatable "mock" json data to be used to ensure a consistent test(s).

Testing:
ran pytest 100% passed
22 passed in 1.00s