Fix incorrect date parsing for Aberdeenshire data
If a number or year appeared in the dataset name then it would be
considered as an incomplete date and populated with the current date's
information to return a complete date. Using the strict parameter will
skip over incomplete dates and return only fully specified information
(which on this page all datasets are dated with).

I did a comparison of the old and new output to ensure that this change
doesn't break any of the currently correct outputs (for example, if any
were only specified to month accuracy).

fixes #108